Safari-609.1.11.1/LayoutTests/performance-api/performance-now-time-origin-in-worker-expected.txt

Ensure time origin in Worker starts from the moment of Worker Creation

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Starting worker: resources/time-origin-in-worker.js
PASS [Worker] performance.now() < 200 is true
PASS successfullyParsed is true

TEST COMPLETE
